pyproject.toml
[project]
name = "InsightBoardReporting"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geopandas>=1.0.1",
    "jinja2>=3.1.4",
    "matplotlib>=3.9.2",
    "numpy>=2.1.1",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pyyaml>=6.0.2",
    "kaleido==0.2.1",
]

[project.scripts]
InsightBoardReporting = "InsightBoardReporting.main:cli"

[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"
